Make Semantic Turkey store the working graph in a graph named after the base uri, instead of utilizing the null context.
This changeset requires that the OWL ART API is compiled from updated sources.

In presence of anomalies as well as to check the correctness of a service, it is possible to use the following SPARQL query to verify that in a Sesame-compliant triple store no statement is asserted in the null-context:

select *
from <http://www.openrdf.org/schema/sesame#nil>
{
?s ?p ?o .
} order by ?s ?p ?o

It could be wise to add a LIMIT clause.