Resolve concurrency hazard in old-style services.
The main effect of the bug was that service requested seemed to mix (e.g. the response to a request could be replaced by the response to a different request).
The cause of the bug is that old-style services are singletons, while the workflow for servicing requests modified the fields of these inherently shared instances, e.g. storing the request inside a field.
The problem was solved by using the class ThreadLocal to have the ability to store different references per thread.

This commit also removed synchronization inside the class XMLHelp by always instantiating new objects for XML processing.