Introduce support for multiple connections in old-style and new-style services.
This support has been implemented as a HandlerInterceptor, which binds a new model to the current
thread before the execution of the handler, and then destroys the model after the handler has executed.

Newer new-style services already supported multiple connections, but that implementation is not compatible
with this. In other words, invoking a newer new-style service does not bind a new OWL ART model to the current
thread.