Fix missing unbinding of a connection from the thread servicing a request.
In case of old-style and new-style services, a HandlerInterceptor is in charge of binding a connection to the
thread servicing a request, and then to unbind the connection. The unbinding is done in the method postHandle,
which seems not be invoked in case of abrupt termination of a service (because of an exception).

Therefore, an exception raised within a service would leave the thread in a corrupted state (manifest via the
exception telling that a model is already bound to the thread).

The solution was to repeat the clean-up in the method afterCompletion, which is guaranteed to be called whatever
the outcome of a service is.