Fix bug in getRoleQueryFragment(..)
The aforementioned method returns a (template of a) graph pattern that can be used for the computation of the role of a resource.
Unfortunately, because of a misspelling in the template, the produced pattern might contain a variable with an erroneous name. The consequence
is that the graph pattern may match other resources than the ones intended.

In particular, we observed that the methods for returning the tree of SKOS collections may contain completely unrelated entries (e.g. owl:import)
because of this bug.