Fix issue ST-294: skos:OrderedCollection should not contain duplicate members
The changeset adds the check before the execution of the service methods for adding elements to ordered
collections, i.e. addFirstToOrderedCollection, addLastToOrderedCollection, addInPositionToOrderedCollection.
Additionally, it introduces a service method for adding elements to unordered collections. While this service
is no more than adding a triple, its presence allows to implement the check. The check may be useless, since
adding an already triple does not change the dataset; however, its presence is motivated by symmetry with the
ordered case.

The check for presence is done (both in the ordered an in the unordered case) by looking at explicit triples
in every graph.