Ignore the datatype rdf:langString inside STRDFLiteralImpl.java
When using an RDF middle adopting the RDF 1.1 semantics for literals, even language tagged literals have a datatype (i.e. rdf:langString). In many places ST makes a strong distinction between typed literals and plain literals. Under the RDF 1.1 semantics many literals are therefore erroneously interpreted as typed literals. A solution to this problem is to ignore the datatype if it is rdf:langString.