Initial implementation of DataRanges inside the resource view (server-side) (ST-585).
The current implementation only produces an appropriate rendering of the datarange, while not providing immediate access to its members (in case of a "one of" datarange). In fact, this is similar to what happens with an enumerated class.

The role dataRange has been changed to be synonym for rdfs:Datatype (as owl:DataRange has been deprecated in OWL 2)