Add the possibility to limit the number of additions/removals in the delta returned by the History service.
The response contains two attributes (additionsTruncated and removalsTruncated) telling the number of triples used for truncating the additions and removals. If the results were not truncated, the corresponding attribute is not present.

When additions/removals are truncated, although the triples are sorted on subject, predicate, object and context, truncation doesn't happen at the end, but arbitrary triples in the middle might be removed.

The UI should indicate the above warning, and possibly use a better expression than truncation.