Fix error in SPARQL-based browsing and introduce an error recovery mode for dealing with rdf:langString literals without an actual language tag.
The solution to the problem with literals was to introduce a custom ValueFactory that sets the language of an rdf:langString literal to "und", if none is provided explicitly. The downside of this approach is that the resource view is showing data that are slightly different from the ones returned by the original provider. However, the problem is somehow mitigated by the fact that this is a read-only visualization.